requirements.txt
Flask==1.1.2
Flask_SQLAlchemy==2.4.4
psycopg2_binary==2.8.6
pytest_flask==1.1.0
gunicorn==20.0.4
flask_restful==0.3.8
webargs==7.0.1
python_dotenv==0.15.0
flask_marshmallow==0.14.0
marshmallow_sqlalchemy==0.24.2